requirements.txt
streamlit==1.30.0
plotly==5.18.0
sentence-transformers==2.6.1
pinecone-client==3.1.0
requests==2.31.0
pandas==2.2.0